feat: add stale support
