+++ package.json
@@ -3,7 +3,7 @@
   "description": "An HTTP compliant route path middleware for serving cache response with invalidation support.",
   "homepage": "https://nicedoc.io/Kikobeats/cacheable-response",
   "version": "2.3.0",
-  "main": "index.js",
+  "main": "src/index.js",
   "author": {
     "email": "josefrancisco.verdu@gmail.com",
     "name": "Kiko Beats",
@@ -55,10 +55,10 @@
   ],
   "dependencies": {
     "@keyvhq/core": "~1.1.1",
+    "@keyvhq/memoize": "~1.2.0",
     "compress-brotli": "~1.3.0",
     "debug-logfmt": "~1.0.4",
     "etag": "~1.8.1",
-    "normalize-url": "~6.1.0",
     "pretty-ms": "~7.0.1"
   },
   "devDependencies": {
@@ -67,6 +67,7 @@
     "ava": "latest",
     "ci-publish": "latest",
     "conventional-github-releaser": "latest",
+    "delay": "latest",
     "finepack": "latest",
     "git-authors-cli": "latest",
     "got": "latest",
@@ -105,8 +106,8 @@
   "license": "MIT",
   "ava": {
     "files": [
-      "!test/util.js",
-      "test/**/*.js"
+      "test/**/*.js",
+      "!test/helpers.js"
     ]
   },
   "commitlint": {
