+++ test/index.js
@@ -1,135 +1,13 @@
 'use strict'
 
-const Keyv = require('@keyvhq/core')
 const test = require('ava')
 const got = require('got')
 
-const { parseCacheControl, createServer } = require('./util')
-
-test('default ttl and revalidate', async t => {
-  const url = await createServer({
-    get: ({ req, res }) => ({ data: { foo: 'bar' } }),
-    send: ({ data, headers, res, req, ...props }) => {
-      res.end('Welcome to Micro')
-    }
-  })
-
-  const { headers } = await got(`${url}/kikobeats`)
-  const cacheControl = parseCacheControl(headers)
-
-  t.true(cacheControl.public)
-  t.true(cacheControl['must-revalidate'])
-  t.true([7199, 7200].includes(cacheControl['max-age']))
-  t.true([1439, 1440].includes(cacheControl['stale-while-revalidate']))
-  t.true([1439, 1440].includes(cacheControl['stale-if-error']))
-})
-
-test('disable revalidation', async t => {
-  const url = await createServer({
-    revalidate: false,
-    get: ({ req, res }) => ({ data: { foo: 'bar' }, ttl: 86400000 }),
-    send: ({ data, headers, res, req, ...props }) => {
-      res.end('Welcome to Micro')
-    }
-  })
-
-  const { headers } = await got(`${url}/kikobeats`)
-  const cacheControl = parseCacheControl(headers)
-
-  t.true(cacheControl.public)
-  t.true(cacheControl['must-revalidate'])
-  t.true([86399, 86400].includes(cacheControl['max-age']))
-  t.false(!!cacheControl['stale-while-revalidate'])
-  t.false(!!cacheControl['stale-if-error'])
-})
-
-test('MISS for first access', async t => {
-  const url = await createServer({
-    get: ({ req, res }) => {
-      return {
-        data: { foo: 'bar' },
-        ttl: 1000,
-        createdAt: Date.now(),
-        foo: { bar: true }
-      }
-    },
-    send: ({ data, headers, res, req, ...props }) => {
-      res.end('Welcome to Micro')
-    }
-  })
-  const { headers } = await got(`${url}/kikobeats`)
-  t.is(headers['x-cache-status'], 'MISS')
-})
-
-test('HIT for second access', async t => {
-  const url = await createServer({
-    get: ({ req, res }) => {
-      return {
-        data: { foo: 'bar' },
-        ttl: 10000,
-        createdAt: Date.now(),
-        foo: { bar: true }
-      }
-    },
-    send: ({ data, headers, res, req, ...props }) => {
-      res.end('Welcome to Micro')
-    }
-  })
-  await got(`${url}/kikobeats`)
-  const { headers } = await got(`${url}/kikobeats`)
-  t.is(headers['x-cache-status'], 'HIT')
-})
-
-test('force query params to invalidate', async t => {
-  const url = await createServer({
-    get: ({ req, res }) => {
-      return {
-        data: { foo: 'bar' },
-        ttl: 86400000,
-        createdAt: Date.now(),
-        foo: { bar: true }
-      }
-    },
-    send: ({ data, headers, res, req, ...props }) => {
-      res.end('Welcome to Micro')
-    }
-  })
-
-  const { headers: headersOne } = await got(`${url}/kikobeats`)
-  t.is(headersOne['x-cache-status'], 'MISS')
-
-  const { headers: headersTwo } = await got(`${url}/kikobeats`)
-  t.is(headersTwo['x-cache-status'], 'HIT')
-
-  const { headers: headersThree } = await got(`${url}/kikobeats?force=true`)
-  t.is(headersThree['x-cache-status'], 'BYPASS')
-  t.is(headersThree['x-cache-expired-at'], '0ms')
-
-  const { headers: headersFour } = await got(`${url}/kikobeats`)
-  t.is(headersFour['x-cache-status'], 'HIT')
-})
-
-test('MISS after cache expiration', async t => {
-  const url = await createServer({
-    get: ({ req, res }) => {
-      return {
-        data: { foo: 'bar' },
-        ttl: 1,
-        createdAt: Date.now(),
-        foo: { bar: true }
-      }
-    },
-    send: ({ data, headers, res, req, ...props }) => {
-      res.end('Welcome to Micro')
-    }
-  })
-  await got(`${url}/kikobeats`)
-  const { headers } = await got(`${url}/kikobeats`)
-  t.is(headers['x-cache-status'], 'MISS')
-})
+const { createServer } = require('./helpers')
 
 test('etag is present', async t => {
   const url = await createServer({
+    staleTtl: false,
     get: ({ req, res }) => {
       return {
         data: { foo: 'bar' },
@@ -211,29 +89,4 @@ test('return empty 304 response when If-None-Match matches ETag', async t => {
   })
   t.is(statusCode, 304)
   t.is(body, '')
-})
-
-test('return HIT after empty 304 response', async t => {
-  const cache = new Keyv({ namespace: 'ssr' })
-  const url = await createServer({
-    cache,
-    get: ({ req, res }) => {
-      return {
-        data: { foo: 'bar' },
-        ttl: 10000,
-        createdAt: Date.now(),
-        foo: { bar: true }
-      }
-    },
-    send: ({ data, headers, res, req, ...props }) => {
-      res.end('Welcome to Micro')
-    }
-  })
-  const { headers: headersOne } = await got(`${url}/kikobeats`)
-  await cache.clear()
-  await got(`${url}/kikobeats`, {
-    headers: { 'If-None-Match': headersOne.etag }
-  })
-  const { headers: headersTwo } = await got(`${url}/kikobeats`)
-  t.is(headersTwo['x-cache-status'], 'HIT')
 })
