+++ test/ttl.js
@@ -1,0 +1,61 @@
+'use strict'
+
+const test = require('ava')
+const got = require('got')
+
+const { parseCacheControl, createServer } = require('./helpers')
+
+test('as value', async t => {
+  const url = await createServer({
+    ttl: 3600000,
+    staleTtl: 720000,
+    get: ({ req, res }) => ({ data: { foo: 'bar' } }),
+    send: ({ data, headers, res, req, ...props }) => {
+      res.end('Welcome to Micro')
+    }
+  })
+
+  const { headers } = await got(`${url}/kikobeats`)
+  const cacheControl = parseCacheControl(headers)
+
+  t.true(cacheControl.public)
+  t.true(cacheControl['must-revalidate'])
+  t.true([3600, 3599].includes(cacheControl['max-age']))
+  t.true([720, 719].includes(cacheControl['stale-while-revalidate']))
+  t.true([720, 719].includes(cacheControl['stale-if-error']))
+})
+
+test('from value', async t => {
+  const url = await createServer({
+    staleTtl: 17280000,
+    get: ({ req, res }) => ({ data: { foo: 'bar' }, ttl: 86400000 }),
+    send: ({ data, headers, res, req, ...props }) => {
+      res.end('Welcome to Micro')
+    }
+  })
+
+  const { headers } = await got(`${url}/kikobeats`)
+  const cacheControl = parseCacheControl(headers)
+
+  t.true(cacheControl.public)
+  t.true(cacheControl['must-revalidate'])
+  t.true([86399, 86400].includes(cacheControl['max-age']))
+  t.true([17279, 17280].includes(cacheControl['stale-while-revalidate']))
+  t.true([17279, 17280].includes(cacheControl['stale-if-error']))
+})
+
+test('sets default ttl', async t => {
+  const url = await createServer({
+    get: ({ req, res }) => ({ data: { foo: 'bar' } }),
+    send: ({ data, headers, res, req, ...props }) => {
+      res.end('Welcome to Micro')
+    }
+  })
+
+  const { headers } = await got(`${url}/kikobeats`)
+  const cacheControl = parseCacheControl(headers)
+
+  t.true(cacheControl.public)
+  t.true(cacheControl['must-revalidate'])
+  t.true([86400, 86399].includes(cacheControl['max-age']))
+})
